+++ CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
  * Add `TypeFactoryTrait::fromValue()` method
  * Deprecate constructing a `CollectionType` instance as a list that is not an array
  * Deprecate the third `$asList` argument of `TypeFactoryTrait::iterable()`, use `TypeFactoryTrait::list()` instead
+ * Add type alias support in `TypeContext` and `StringTypeResolver`
 
 7.2
 ---